services/trainer/requirements.txt
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
joblib==1.4.2
psycopg2-binary==2.9.9
